Move class methods under __init__() method
Per python documentation: "When a class defines an __init__() method, class instantiation automatically invokes __init__() for the newly-created class instance."

In other words, it is the first method called by the class when an instance of it is initialized/when the class is called. It's usually best to list functions with items that run first listed first so it reads similar to how it runs. (Except for the case of something needing to be defined first before used such as functions and classes; the body of the function/class do not run until they are called but you have to list then in the program before you can call them, so the function/class name is in the namespace.)