Make ProgramBody subclass of ttk.Frame. Add Replay
This adjustment makes it somewhat easier to create widgets. It also makes more sense in the grand scheme of the class, since it was effectively a subclass of ttk.Frame but did it more manually within __init__ (and _create_widgets()). This makes it more clear that ProgramBody is meant to be a class that is essentially a Tkinter Frame object.

Didn't realize I accidentally removed the Play Again Button. Swapped the Exit button location with Play Again since it seemed to make more sense to have Exit at the end.