Add if __name__ == '__main__'
Not necessarily needed, but if this was ever made into a package, this makes it so that the Chordial application will not unintentionally when if someone puts `import Chordial` into another python file. __name__ will only equal to '__main__' if file is run directly through VSCode, the terminal with python.exe Chrodial.py, etc.