Replace Exit button with Remove from Suggestions
Tkinter and script can be exited by closing the window, making the button unnecessary. Also removed to make room for Remove from Suggestions button.

I thought it would be convenient to have a button to remove a song from the suggestions if you do not want it to be selected again by Chordial. 

This button removes the suggestion from the list within the running script and does not do anything to the actual files. If Chordial is closed and opened again, the sound that was removed will be back in the list if the user didn't dislike that sound so much that they decided to delete it from their mid_directory.